Refactored courtbotError into its own module
* refactored courtbotError into its own module
* moved courtbotError unit tests into own test script
