+++ src/events.js
@@ -1,4 +1,6 @@
 const EventEmitter = require(`events`);
+import courtbotError from '../src/courtbotError'
+import {COURTBOT_ERROR_NAME} from '../src/courtbotError';
 
 class CourtbotEmitter extends EventEmitter {};
 
@@ -125,27 +127,4 @@ export function getCasePartyEvents(casenumber, party, errorMode = 1) {
     }
     return results;
   });
-}
-
-// A well-formed custom error class for logging/tracing purposes
-// See https://www.bennadel.com/blog/2828-creating-custom-error-objects-in-node-js-with-error-capturestacktrace.htm
-const COURTBOT_ERROR_NAME = `Courtbot Error`;
-
-export class courtbotError extends Error {
-  constructor(settings = {}, context) {
-    super();
-
-    settings = settings || {};
-
-    this.name = COURTBOT_ERROR_NAME;
-    this.type = settings.type || `general`;
-    this.message = settings.message || `No message listed`;
-    this.case = settings.case || `No case listed`;
-    this.api = settings.api || `No api listed`;
-    this.timestamp = settings.timestamp || `No timestamp listed`;
-    this.initialError = settings.initialError || null;
-    this.isCourtbotError = true; // undefined values are falsy
-
-    Error.captureStackTrace(this, (context || courtbotError));
-  }
 }
